Add serialization support to the model and integrate the output (JSON) into the status page so that the content of the model can be viewed via the API handler.
Refactor the Eagle interface effectively removing it.  By making it a concrete type, the inclusion of functionality doesn't look weird.  The Eagle statistics tracking has been moved out into its own type.

An initial /status route test case has been added however it doesn't *actually* validate that the route works.  It does, however, cause the route to generate a response.  The unit tests are still inconsistent with respect to how the operator<< for custom types is added for BOOST_EQUAL.

Add support for a custom Boost binary (built for AMD64) as the Ubuntu 20.04 (or 21.xx) doesn't support > 1.74 and thus doesn't contain the Boost.Json dependency which appeared in 1.75.  Additionally, add scripts to the deps which enables the integration of clang-12 for container build purposes.

Finally, the Howard Hinnant date repo has been added as a dependency submodule.  This enables the translation of std::chrono::time_point into human-readable dates; this will appear in C++20 but is not in C++17.