Added the COPYING file containing the license for this project.
Added to two missing fragment processors (although these are undocumented in the spec.  There is some basic processing but it is likely incorrect.

Fixed some bugs around the use of string_view where the original string goes out of scope.  Refactored some of the property_tree type wranglers into the original classes for said types.

Added a new "IsExpected" to complement "IsEssential" and "IsOptional"; it's behaviour is effectively to loudly complain about lack of spec conformity (i.e. non-optional field is missing) but to allow continued processing as the field is not used for anything sensitive.