Refactoring of how optional and essential field processors are handled (so that the fragment processing is in a single place).
Added "optional" support to each of the fields in the various fragment types where that field is not strictly required for a calculation or immediate usage.  Optional fields may be missing from a fragment and not cause significant consequences.

Cleaned up some legacy class types and functions which were no longer used.